requirements.txt
functions-framework
python-dotenv
sendgrid